Fixing data frame index issue
If the data frame doesn't have an index == 0 it will raise an error.

Running the command pyreadr.write_rdata for the whole data frame will probably work but if the data frame is filtered it raise. This can't be avoided if a df.reset_index() is performed beforehand but could take some time until the developer discovers the issue.